+++ system/modules/core/library/Contao/Combiner.php
@@ -325,11 +325,12 @@ class Combiner extends \System
 		if ($arrFile['extension'] == self::SCSS)
 		{
 			$objCompiler = new \scssc();
+			new \scss_compass($objCompiler);
 
 			$objCompiler->setImportPaths(array
 			(
 				TL_ROOT . '/' . dirname($arrFile['name']),
-				TL_ROOT . '/assets/compass/' . $GLOBALS['TL_ASSETS']['COMPASS']
+				TL_ROOT . '/vendor/leafo/scssphp-compass/stylesheets'
 			));
 
 			$objCompiler->setFormatter((\Config::get('debugMode') ? 'scss_formatter' : 'scss_formatter_compressed'));