+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.3.2 (2014-XX-XX)
 --------------------------
 
 ### Fixed
+Render all root nodes in the page and file picker (see #6844).
+
+### Fixed
 Add the "scssphp-compass" library to support Compass functions.
 
 ### Fixed